1.混淆更新

+++ plugin/src/main/java/top/niunaijun/blackobfuscator/core/ObfDex.java
@@ -13,22 +13,14 @@ import java.util.List;
 
 import top.niunaijun.obfuscator.ObfuscatorConfiguration;
 
-/**
- * Created by Milk on 2021/12/17.
- * * ∧＿∧
- * (`･ω･∥
- * 丶　つ０
- * しーＪ
- * 此处无Bug
- */
 public class ObfDex {
+
     public static void obf(String dir, int depth, String[] obfClass, String[] blackClass, String mappingFile) {
         File file = new File(dir);
         Mapping mapping = new Mapping(mappingFile);
         if (file.isDirectory()) {
             File[] files = file.listFiles();
-            if (files == null)
-                return;
+            if (files == null) return;
             for (File input : files) {
                 if (input.isFile()) {
                     handleDex(input, depth, obfClass, blackClass, mapping);
@@ -42,18 +34,22 @@ public class ObfDex {
     }
 
     private static void handleDex(File input, int depth, String[] obfClass, String[] blackClass, Mapping mapping) {
-        if (!input.getAbsolutePath().endsWith(".dex"))
-            return;
+        if (!input.getAbsolutePath().endsWith(".dex")) return;
         File tempJar = null;
         File splitDex = null;
         File obfDex = null;
+        File new_tempJar = null;
+
         try {
             tempJar = new File(input.getParent(), System.currentTimeMillis() + "obf" + input.getName() + ".jar");
             splitDex = new File(input.getParent(), System.currentTimeMillis() + "split" + input.getName() + ".dex");
             obfDex = new File(input.getParent(), System.currentTimeMillis() + "obf" + input.getName() + ".dex");
+            new_tempJar = new File(tempJar.getParent(), "new_" + tempJar.getName());
+
             List<String> obfClassList = arrayToList(obfClass);
             List<String> blackClassList = arrayToList(blackClass);
 
+            // 解析官方的类名混淆文件 获取白名单中被映射后的类名
             TrieTree whiteListTree = new TrieTree();
             whiteListTree.addAll(obfClassList);
 
@@ -67,6 +63,7 @@ public class ObfDex {
                 }
             }
 
+            // 解析官方的类名混淆文件 获取黑名单中被映射后的类名
             TrieTree blackListTree = new TrieTree();
             blackListTree.addAll(blackClassList);
             List<String> tmpBlackClass = new ArrayList<>(blackClassList);
@@ -79,12 +76,17 @@ public class ObfDex {
                     }
                 }
             }
+
+            // 通过 dexLib2 解析出在 obfClassList 在 blackClassList 中的类，分别编译成 smali 文件，再合并成新的 dex 文件
+            // 并输出到 splitDex 指定的路径
             long l = DexLib2Utils.splitDex(input, splitDex, obfClassList, blackClassList);
             if (l <= 0) {
                 System.out.println("Obfuscator Class not found");
                 return;
             }
 
+            // 使用 dex2jar 将 .dex 转换成 .class(.jar)
+            // 具体就是把 splitDex -> tempJar
             new Dex2jarCmd(new ObfuscatorConfiguration() {
                 @Override
                 public int getObfDepth() {
@@ -97,14 +99,28 @@ public class ObfDex {
                     return super.accept(className, methodName);
                 }
             }).doMain("-f", splitDex.getAbsolutePath(), "-o", tempJar.getAbsolutePath());
-            new Jar2Dex().doMain("-f", "-o", obfDex.getAbsolutePath(), tempJar.getAbsolutePath());
+
+            //            new Jar2Dex().doMain("-f", "-o", obfDex.getAbsolutePath(), tempJar.getAbsolutePath());
+
+            // 字符串混淆 string -> byte[] 异或 + Base64 异或因子随机，内联解密以防 hook
+            Operator.run(tempJar, new_tempJar, true);
+            new Jar2Dex().doMain("-f", "-o", obfDex.getAbsolutePath(), new_tempJar.getAbsolutePath());
             DexLib2Utils.mergerAndCoverDexFile(input, obfDex, input);
         } catch (Throwable t) {
             t.printStackTrace();
         } finally {
-            tempJar.delete();
-            splitDex.delete();
-            obfDex.delete();
+            if (tempJar != null) {
+                tempJar.delete();
+            }
+            if (new_tempJar != null) {
+                new_tempJar.delete();
+            }
+            if (splitDex != null) {
+                splitDex.delete();
+            }
+            if (obfDex != null) {
+                obfDex.delete();
+            }
         }
     }
